refactor: Remove circular dependency
