+++ app/configurator/components/field.tsx
@@ -78,10 +78,10 @@ import {
   Component,
   Dimension,
   HierarchyValue,
-  ObservationValue,
   TemporalDimension,
   isTemporalOrdinalDimension,
 } from "@/domain/data";
+import { VISUALIZE_MAX_VALUE, isDynamicMaxValue } from "@/domain/max-value";
 import { useTimeFormatLocale } from "@/formatters";
 import { TimeUnit } from "@/graphql/query-hooks";
 import { useLocale } from "@/locales/use-locale";
@@ -310,21 +310,6 @@ export const DataFilterSelect = ({
       value={usesMostRecentValue ? maxValue : fieldProps.value}
     />
   );
-};
-
-/** We can pin some filters' values to max value dynamically, so that when a new
- * value is added to the dataset, it will be automatically used as default filter
- * value for published charts.
- */
-export const VISUALIZE_MAX_VALUE = "VISUALIZE_MAX_VALUE";
-
-/** Checks if a given filter value is supposed to be dynamiaclly pinned to max
- * value.
- */
-export const isDynamicMaxValue = (
-  value: ObservationValue
-): value is "VISUALIZE_MAX_VALUE" => {
-  return value === VISUALIZE_MAX_VALUE;
 };
 
 type DataFilterTemporalProps = {
